# functions/swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Bet Chicago
host: localhost:5001
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
  - application/vnd.contentful.management.v1+json
  - application/x-www-form-urlencoded
  - multipart/form-data
  - text/plain; charset=utf-8
  - image/jpeg
produces:
  - application/json
paths:
...
  /golf/tournament:
    x-swagger-router-controller: golf
    post:
      description: update firedata from sports radar
      operationId: tournament
      tags:
        - Golf
      parameters:
        - name: tournamentId
          in: query
          description: id of tournament :applicationConfig.golfTournament is default
          required: false
          type: string
        - name: year
          in: query
          description: year :application.golfYearPGA is default
          required: false
          type: string
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /odds/updateodds:
    x-swagger-router-controller: odds
    post:
      description: update odds (for current day)
      operationId: updateOdds
      tags:
        - Odds
      parameters:
        - name: date
          in: query
          description: date if not current (current is best)
          type: string
          required: false
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /odds/udpateschedule:
    x-swagger-router-controller: odds
    post:
      description: update schedule information for a sport / date
      operationId: updateODDSchedule
      tags:
        - Odds
      parameters:
        - name: sportid
          in: query
          description: sr id (sport radar)
          type: string
          required: true
        - name: date
          in: query
          description: date if not current
          type: string
          required: false
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...
  /odds/udpatetournamentsbysport:
    x-swagger-router-controller: odds
    post:
      description: update tournaments for a sport
      operationId: updateTournamentsBySport
      tags:
        - Odds
      parameters:
        - name: sportid
          in: query
          description: sr id (sport radar)
          type: string
          required: true
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /odds/updateoddsperformance:
    x-swagger-router-controller: odds
    post:
      description: process odds data and stats data into spreadPerformance
      operationId: updateOddsPerformance
      tags:
        - Odds
      parameters:
        - name: sportkey
          in: query
          description: prefix in database to daily info (i.e. ncaamb)
          type: string
          required: true
        - name: oddskey
          in: query
          description: odds data category (i.e. Basketball)
          type: string
          required: true
        - name: date
          in: query
          description: date if not current
          type: string
          required: false
        - name: dateoffset
          type: string
          in: query
          description: day ahead or behind to process (-1 is yesterday)
          required: false
      responses:
        "200":
          description: job completion status
          schema:
            type: string
  /odds/updateatshistory:
    x-swagger-router-controller: odds
    post:
      description: process odds data and stats data into spreadPerformance
      operationId: updateATSHistory
      tags:
        - Odds
      parameters:
        - name: sportkey
          in: query
          description: prefix in database to daily info (i.e. ncaamb)
          type: string
          required: true
      responses:
        "200":
          description: job completion status
          schema:
            type: string
...